question access control

+++ src/main/java/by/latushko/anyqueries/controller/command/impl/post/DeleteQuestionCommand.java
@@ -13,11 +13,13 @@ import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import jakarta.servlet.http.HttpSession;
 
+import static by.latushko.anyqueries.controller.command.CommandResult.RoutingType.FORWARD;
 import static by.latushko.anyqueries.controller.command.CommandResult.RoutingType.REDIRECT;
 import static by.latushko.anyqueries.controller.command.ResponseMessage.Level.DANGER;
 import static by.latushko.anyqueries.controller.command.ResponseMessage.Level.SUCCESS;
 import static by.latushko.anyqueries.controller.command.identity.CookieName.LANG;
 import static by.latushko.anyqueries.controller.command.identity.HeaderName.REFERER;
+import static by.latushko.anyqueries.controller.command.identity.PagePath.ERROR_403_PAGE;
 import static by.latushko.anyqueries.controller.command.identity.PageUrl.QUESTIONS_URL;
 import static by.latushko.anyqueries.controller.command.identity.RequestParameter.ID;
 import static by.latushko.anyqueries.controller.command.identity.RequestParameter.PAGE;
@@ -30,17 +32,20 @@ public class DeleteQuestionCommand implements Command {
     @Override
     public CommandResult execute(HttpServletRequest request, HttpServletResponse response) {
         HttpSession session = request.getSession();
+        User user = (User) session.getAttribute(PRINCIPAL);
+        Long id = getLongParameter(request, ID);
+        QuestionService questionService = QuestionServiceImpl.getInstance();
+        if(!questionService.checkDeleteAccess(id, user)) {
+            return new CommandResult(ERROR_403_PAGE, FORWARD);
+        }
         String referer = QueryParameterHelper.removeParameter(request.getHeader(REFERER), PAGE);
         if(!referer.contains(QUESTIONS_URL)) {
             referer = QUESTIONS_URL;
         }
         CommandResult commandResult = new CommandResult(referer, REDIRECT);
-        Long id = getLongParameter(request, ID);
         String userLang = CookieHelper.readCookie(request, LANG);
         MessageManager manager = MessageManager.getManager(userLang);
         ResponseMessage message;
-        QuestionService questionService = QuestionServiceImpl.getInstance();
-        User user = (User) session.getAttribute(PRINCIPAL);
         boolean result = questionService.delete(id, user);
         if(result) {
             message = new ResponseMessage(SUCCESS, manager.getMessage(MESSAGE_DELETE_SUCCESSFUL));
